// src/app/interfaces/sesion.ts
export interface Sesion{
    idsesion: number;
    idusuario: number;
    nombre: string;
    apellidos: string;
    telefono: string;
    email: string;
    contrasena: string;
    domicilio: string;
    tipo: string;
}

export interface sesion{
    idsesion: number;
    idusuario: number;
    nombre: string;
    apellidos: string;
    telefono: string;
    email: string;
    contrasena: string;
    domicilio: string;
    tipo: string;
}